Refactored user related code
g.user now is an instance of UserModel.
