+++ schema/query.py
@@ -1,3 +1,6 @@
+import graphene
+from graphene import Field, List, String
+from flask import g
 from schema.partner import *
 from schema.selectors import Selectors
 from schema.proposal import *
@@ -7,9 +10,6 @@ from data.targets import get_targets
 from data.selectors import get_selectors_data
 from schema.instruments import *
 from schema.user import UserModel
-from data.user import get_user
-import graphene
-from graphene import Field, List, String
 
 
 class Query(graphene.ObjectType):
@@ -56,7 +56,7 @@ class Query(graphene.ObjectType):
         return get_selectors_data()
 
     def resolve_user(self, context, info, args):
-        return get_user()
+        return g.user
 
 
 schema = graphene.Schema(query=Query, types=[HRS, RSS, BVIT, SCAM])
